Desktop/POS/Business Dashboard/requirements.txt
flask==2.3.3
flask-socketio==5.3.6
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
eventlet==0.33.3
plotly==5.18.0
pandas==2.1.1
openai==1.3.5
python-dotenv==1.0.0 